variable correction

+++ includes/duplicates.php
@@ -5,34 +5,42 @@ class duplicates {
 	public static function updateDupeTable($formID,$objectID,$data) {
 
 		// trans begin
+		$result = mfcs::$engine->openDB->transBegin("objects");
+		if ($result !== TRUE) {
+			errorHandle::newError(__METHOD__."() - unable to start database transactions", errorHandle::DEBUG);
+			return FALSE;
+		}
 
 		// wipe the old dupe information
 		$sql       = sprintf("DELETE FROM `dupeMatching` WHERE `formID`='%s' AND `objectID`='%s'",
-			$engine->openDB->escape($formID),
-			$engine->openDB->escape($objectID)
+			mfcs::$engine->openDB->escape($formID),
+			mfcs::$engine->openDB->escape($objectID)
 			);
-		$sqlResult = $engine->openDB->query($sql);
+		$sqlResult = mfcs::$engine->openDB->query($sql);
 
 		if (!$sqlResult['result']) {
-			$engine->openDB->transRollback();
-			$engine->openDB->transEnd();
+			mfcs::$engine->openDB->transRollback();
+			mfcs::$engine->openDB->transEnd();
 			errorHandle::newError(__METHOD__."() - removing from duplicate table: ".$sqlResult['error'], errorHandle::DEBUG);
 			return FALSE;
 		}
 
 		//insert data
 		foreach ($data as $name=>$raw) {
+
+			if (!isset(mfcs::$engine->cleanPost['MYSQL'][$name]) || isempty(mfcs::$engine->cleanPost['MYSQL'][$name])) continue;
+
 			$sql       = sprintf("INSERT INTO `dupeMatching` (`formID`,`objectID`,`field`,`value`) VALUES('%s','%s','%s','%s')",
-				$engine->openDB->escape($formID),
-				$engine->openDB->escape($objectID),
-				$engine->openDB->escape($name),
-				$engine->cleanPost['MYSQL'][$name]
+				mfcs::$engine->openDB->escape($formID),
+				mfcs::$engine->openDB->escape($objectID),
+				mfcs::$engine->openDB->escape($name),
+				mfcs::$engine->cleanPost['MYSQL'][$name]
 			);
-			$sqlResult = $engine->openDB->query($sql);
+			$sqlResult = mfcs::$engine->openDB->query($sql);
 
 			if (!$sqlResult['result']) {
-				$engine->openDB->transRollback();
-				$engine->openDB->transEnd();
+				mfcs::$engine->openDB->transRollback();
+				mfcs::$engine->openDB->transEnd();
 
 				errorHandle::newError(__METHOD__."() - : ".$sqlResult['error'], errorHandle::DEBUG);
 				return FALSE;
@@ -40,8 +48,8 @@ class duplicates {
 		}
 
 		// trans commit
-		$engine->openDB->transCommit();
-		$engine->openDB->transEnd();
+		mfcs::$engine->openDB->transCommit();
+		mfcs::$engine->openDB->transEnd();
 
 		return TRUE;
 	}
